inicio de commit

+++ src/main/java/com/ibs/proyecto/repository/ICategoriaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ibs.proyecto.model.Categoria;
 
-public interface ICategoriaRepository extends CrudRepository<Categoria, Integer>{
+public interface ICategoriaRepository extends CrudRepository<Categoria, Long>{
 
 }
