+++ src/main/java/com/ibs/proyecto/repository/IClienteRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ibs.proyecto.model.Cliente;
 
-public interface IClienteRepository extends CrudRepository<Cliente, Integer>{
+public interface IClienteRepository extends CrudRepository<Cliente, Long>{
+
 
 }
