+++ src/main/java/com/ibs/proyecto/repository/IMarcaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ibs.proyecto.model.Marca;
 
-public interface IMarcaRepository extends CrudRepository<Marca, Integer>{
+public interface IMarcaRepository extends CrudRepository<Marca, Long>{
 
 }
