+++ src/main/java/com/ibs/proyecto/repository/IUsuarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ibs.proyecto.model.Usuario;
 
-public interface IUsuarioRepository extends CrudRepository<Usuario, Integer>{
+public interface IUsuarioRepository extends CrudRepository<Usuario, Long>{
 
 }
